Fix unicode font issue for VTerm
Vterm was not showing the correct characters in the ZSH Powerline based
prompt.  The issue was that the specified font i.e. Iosevka Nerd Font,
was not being used for these characters.  The fix was to inform Doom
that it should use Isoveka Nerd Font for the unicode font as well.