Remove todo stats roll up
Used this for awhile where I wanted high level headings to roll up all
the TODOs and checkbox items.  However this ends up effectively making
each sub-task the same weight as every other one.  So moving to not
doing the recursive roll-up so that parent headings give a roll-up of
just immediate children items, which is a closer reflection of
completion given we can put children with the same weighting under each
parent heading.