Adds to org-agenda-log-mode-items: state
In the Org agenda log mode view the default shows when items are closed
and clocked-in. This change adds an entry for state changes, which is
useful to see when items are being started.. or even put on hold, etc.
The clock state is not flexible enough on its own as it only allows one
task to be reflected at a time.